requirements.txt
torch~=1.13.1
opencv-python~=4.9.0.80
numpy~=1.21.6
Pillow~=9.5.0
torchvision~=0.14.1
pandas~=1.3.5
openpyxl~=3.1.3
tqdm~=4.66.4
requests~=2.31.0
paramiko~=3.4.0